Extract ContextualInfo component
