+++ ts/screens/authentication/LandingScreen.tsx
@@ -13,7 +13,7 @@ import { NavigationInjectedProps } from "react-navigation";
 import { connect } from "react-redux";
 import ButtonDefaultOpacity from "../../components/ButtonDefaultOpacity";
 import CieNotSupported from "../../components/cie/CieNotSupported";
-import { ContextualHelp } from "../../components/ContextualHelp";
+import { ContextualHelp } from "../../components/ContextualInfo";
 import { DevScreenButton } from "../../components/DevScreenButton";
 import { withLightModalContext } from "../../components/helpers/withLightModalContext";
 import { HorizontalScroll } from "../../components/HorizontalScroll";
