+++ ts/screens/profile/ProfileMainScreen.tsx
@@ -12,7 +12,7 @@ import {
 import { connect } from "react-redux";
 import { TranslationKeys } from "../../../locales/locales";
 import ButtonDefaultOpacity from "../../components/ButtonDefaultOpacity";
-import { ContextualHelp } from "../../components/ContextualHelp";
+import { ContextualHelp } from "../../components/ContextualInfo";
 import FiscalCodeComponent from "../../components/FiscalCodeComponent";
 import { withLightModalContext } from "../../components/helpers/withLightModalContext";
 import { ContextualHelpPropsMarkdown } from "../../components/screens/BaseScreenComponent";
